+++ CHANGELOG.md
@@ -4,11 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [v0.2.0]
 ### Added
 - Coverage plots to alignment stats report
 
 ### Changed
+- Improved documentation and new structure of the output sub-directories
 - Variant allele frequency representation is now a scatterplot showing the relationship between normal and tumor VAF
 - Created a subworkflow to call somatic SV (`somatic_sv` in `workflows/wf-somatic-sv.nf`)
 - Add nanomonsv soft filtering SV when providing a bed file specifying the tandem repeat with `--tr_bed`